% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicater-package.R
\docType{package}
\name{replicater-package}
\alias{replicater}
\alias{replicater-package}
\title{replicater: Replicate multiple scripts and issue a report}
\description{
Thids package intends to replicate multiple scripts contained, and issue a tabular odcument reporting the errors encountered.
}
\author{
\strong{Maintainer}: First Last \email{Matthieu.Stigler@example.com}

}
\keyword{internal}
