% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{rpl_check_deps}
\alias{rpl_check_deps}
\title{Check dependencies in package}
\usage{
rpl_check_deps(path_scripts = ".")
}
\arguments{
\item{path_scripts}{path where the scripts are located}
}
\description{
Check dependencies in package
}
