% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicater-package.R
\docType{package}
\name{replicater-package}
\alias{replicater}
\alias{replicater-package}
\title{Some description}
\description{
This package provides tools for replicating R scripts. It is stored in: https://github.com/MatthieuStigler/replicater.
}
\author{
\strong{Maintainer}: Matthieu Stigler \email{Matthieu.Stigler@example.com}

}
\keyword{internal}
